Use scoped enums (enum class) instead of unscoped enums
* examples/dom_build/main.cc:
* examples/dom_xpath/main.cc: Update names of enum constants.
* libxml++/document.[cc|h]:
* libxml++/nodes/node.[cc|h]:
* libxml++/parsers/parser.[cc|h]:
* libxml++/parsers/textreader.[cc|h]: Replace enum by enum class.
Modify the names of the enum constants. E.g. XML_INTERNAL_GENERAL_ENTITY ->
XmlEntityType::INTERNAL_GENERAL.

This patch breaks API and ABI. The API and ABI of libxml++ 3.x has not yet
been frozen.

+++ libxml++/document.cc
@@ -415,10 +415,11 @@ void Document::set_entity_declaration(const Glib::ustring& name, XmlEntityType t
                               const Glib::ustring& publicId, const Glib::ustring& systemId,
                               const Glib::ustring& content)
 {
-  auto entity = xmlAddDocEntity( impl_, (const xmlChar*) name.c_str(), type,
+  auto entity = xmlAddDocEntity(impl_, (const xmlChar*)name.c_str(),
+    static_cast<int>(type),
     publicId.empty() ? nullptr : (const xmlChar*)publicId.c_str(),
     systemId.empty() ? nullptr : (const xmlChar*)systemId.c_str(),
-    (const xmlChar*) content.c_str() );
+    (const xmlChar*)content.c_str());
   if (!entity)
     throw internal_error("Could not add entity declaration " + name);
 }
